fixes #1968 fixes #1967 fixes #1959
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@11762 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/gui/swing/dialog/CaptchaDialog.java
@@ -96,7 +96,7 @@ public class CaptchaDialog extends JCountdownDialog implements ActionListener, K
 
         int size = SubConfiguration.getConfig("JAC").getIntegerProperty(Configuration.PARAM_CAPTCHA_SIZE, 100);
         if (size != 100) {
-            imageIcon = new ImageIcon(imageIcon.getImage().getScaledInstance(imageIcon.getIconWidth() * (size / 100), imageIcon.getIconHeight() * (size / 100), Image.SCALE_SMOOTH));
+            imageIcon = new ImageIcon(imageIcon.getImage().getScaledInstance((int) (imageIcon.getIconWidth() * (size / 100.0f)), (int) (imageIcon.getIconHeight() * (size / 100.0f)), Image.SCALE_SMOOTH));
         }
 
         textField = new JTextField(10);
